separate out HaproxyManager
