+++ src/reload.go
@@ -18,10 +18,6 @@ import (
 
 	nplus "github.com/nginxinc/nginx-plus-go-client/client"
 	"github.com/sirupsen/logrus"
-
-	runtime_models "github.com/haproxytech/client-native/v5/models"
-	runtime_api "github.com/haproxytech/client-native/v5/runtime"
-	runtime_options "github.com/haproxytech/client-native/v5/runtime/options"
 )
 
 type NamespaceRenderingData struct {
@@ -152,7 +148,12 @@ func reload() error {
 			// Create a context with a timeout for the API call.
 			ctx, cancel := context.WithTimeout(context.Background(), time.Duration(config.apiTimeout)*time.Second)
 			defer cancel()
-			err = haproxyRuntimeAPI(&data, ctx)
+			haproxyMgr, mgrErr := NewHaproxyManager(ctx, config.HaproxySocketAddr, config.HaproxyDisableLargeBackendCountOptimisation)
+			if mgrErr != nil {
+				err = mgrErr
+			} else {
+				err = haproxyMgr.ReconcileAllBackends(&data, config.HaproxyDisableLargeBackendCountOptimisation)
+			}
 		}
 
 		// Update health status based on the result of the API call
@@ -403,34 +404,6 @@ func IsUnixSocketAddr(addr string) bool {
 	return true
 }
 
-func newHAProxyClient(ctx context.Context) (runtime_api.Runtime, error) {
-	haproxySocket := config.HaproxySocketAddr
-	logger.WithField("haproxy_socket", haproxySocket).Debug("Preparing to connect to HAProxy runtime API")
-
-	if haproxySocket == "" {
-		return nil, errors.New("HAProxy socket address is not configured")
-	}
-
-	if IsUnixSocketAddr(haproxySocket) {
-		if _, err := os.Stat(haproxySocket); os.IsNotExist(err) {
-			return nil, fmt.Errorf("HAProxy socket file does not exist: %s", haproxySocket)
-		}
-	}
-
-	runtimeClient, err := runtime_api.New(ctx, runtime_options.Sockets(map[int]string{1: haproxySocket}))
-	if err != nil {
-		return nil, fmt.Errorf("error initializing HAProxy runtime client: %w", err)
-	}
-
-	// Test the connection to ensure the runtime API is responsive.
-	if _, err := runtimeClient.GetInfo(); err != nil {
-		return nil, fmt.Errorf("error connecting to HAProxy socket: %w", err)
-	}
-
-	logger.Info("Successfully connected to HAProxy runtime API")
-	return runtimeClient, nil
-}
-
 func isHTTPHostGroup(hosts []Host) bool {
 	for _, host := range hosts {
 		if host.PortType != "http" && host.PortType != "https" {
@@ -438,366 +411,6 @@ func isHTTPHostGroup(hosts []Host) bool {
 		}
 	}
 	return true
-}
-
-
-func haproxyRuntimeAPI(data *RenderingData, ctx context.Context) error {
-	start := time.Now()
-	var resultLabel string
-	defer func() {
-		duration := time.Since(start)
-		haproxyReconcileAllBackendsDuration.WithLabelValues(resultLabel).Observe(duration.Seconds())
-		statsTimingVec("haproxy_runtime_api_duration", duration, resultLabel)
-	}()
-
-	config.RLock()
-	defer config.RUnlock()
-
-	runtimeClient, err := newHAProxyClient(ctx)
-	if err != nil {
-		resultLabel = "error"
-		updateHealthForUpstreamUpdateAPI(false, err.Error())
-		return err
-	}
-
-	// Aggregate all desired hosts for each unique backend from our configuration. We don't want to make excess API calls
-	backendsToReconcile := make(map[string][]Host)
-	for _, app := range data.Apps {
-		if app.Vhost == "" {
-			continue
-		}
-		for groupName, groupData := range app.Groups {
-			if len(groupData.Hosts) == 0 {
-				continue
-			}
-
-			if !isHTTPHostGroup(groupData.Hosts) {
-				logger.WithFields(logrus.Fields{
-					"Vhost": app.Vhost,
-					"group": groupName,
-					"hosts": groupData.Hosts,
-				}).Warning("Non-HTTP/HTTPS host group detected. Skipping HAProxy runtime API update for this group.")
-				continue
-			}
-
-			backendName := generateStableHaproxyBackendName(app, groupName)
-			if backendName != "" {
-				backendsToReconcile[backendName] = append(backendsToReconcile[backendName], groupData.Hosts...)
-			}
-
-		}
-	}
-
-	// Reconcile each unique backend. GetServersState (i.e. show servers state %backendName ) can only be called once per backend. Runtime API library doesn't return backend names
-	// when calling GetServersState as models.RuntimeServers is a list of []*RuntimeServer without backend context returned by models.ParseRuntimeServer()
-	// If number of backends is large, this could take a while and instead we will use equivalent of "show servers state <all>" which returns all backends and servers in one call
-	// we will only call GetServersState if we see a diff in server names expected for a backend
-
-	//For clusters with small number of backends, calling GetServersState for each backend is not a big overhead
-	//but for a lot of backends, this can take a lot of time and we could have stale data for backends processed later in the loop
-	//E.g. For a heavily loaded HAProxy with 1000 backends and >4000 RPS, calling GetServersState for each backend could take upwards of 200 seconds
-	//So we will call GetServersState only if we don't find the backend in aggregated call to our own implementation of GetServersStateWithBackend
-	//This way we optimize for both small and large number of backends
-
-	//Upstream runtime doesn't expose the backend name when calling GetServersState API
-	//so we have our own implementation which calls "show servers state" command and parses the output to get servers grouped by backend name
-
-	//We could also use GetStats API but that would require more processing to convert stats to server state
-	//and stats don't show servers in maintenance mode unless they are enabled. So a disabled server in maintenance mode
-	//would not be visible in stats output but it is visible in "show servers state" output
-
-	//Despite the fact that "show servers state" is reliable, making the use configurable in case of any issues of changes to models in future versions of HAProxy client runtime
-	logger.WithField("count", len(backendsToReconcile)).Debug("Reconciling all unique HAProxy backends")
-	reconciledBackends := make(map[string]bool)
-	reconciliationFailedBackends := make(map[string]bool)
-	allServersState := make(map[string]runtime_models.RuntimeServers)
-	if !config.HaproxyDisableLargeBackendCountOptimisation {
-		logger.Debug("HAProxy large backend count optimisation is enabled. Using aggregated call to get all backends and servers state")
-		allServersState, err = haproxyClientGetServersStateWithBackend(runtimeClient)
-		if err != nil {
-			resultLabel = "error"
-			updateHealthForUpstreamUpdateAPI(false, err.Error())
-			logger.WithFields(logrus.Fields{
-				"error": err,
-			}).Error("Failed to get HAProxy servers state for all backends")
-			haproxyAPICallsFailed.WithLabelValues("get_servers_state_all_backends").Inc()
-			statsCountVec("haproxy_api_calls_failed_total", 1, "get_servers_state_all_backends")
-			return fmt.Errorf("failed to get HAProxy servers state for all backends: %w", err)
-		}
-		haproxyAPICallsSuccessful.WithLabelValues("get_servers_state_all_backends").Inc()
-		statsCountVec("haproxy_api_calls_successful_total", 1, "get_servers_state_all_backends")
-	}
-
-	for backend, hosts := range backendsToReconcile {
-
-		currentServersForBackend := []*runtime_models.RuntimeServer{}
-		err := error(nil)
-		// If we have state for this backend from the aggregated call, use it directly.
-		// This avoids making an additional API call per backend.
-		if !config.HaproxyDisableLargeBackendCountOptimisation || allServersState[backend] != nil {
-			currentServersForBackend = allServersState[backend]
-		} else {
-			if allServersState[backend] != nil {
-				logger.Warn("Consider disabling large backend count optimisation with haproxy_disable_large_backend_count_optimisation set to true")
-			}
-			logger.WithField("backend", backend).Debug("No existing servers found for backend in aggregated state. Trying to get state for the backend directly")
-			currentServersForBackend, err = runtimeClient.GetServersState(backend)
-		}
-
-		if err != nil {
-			reconciliationFailedBackends[backend] = true
-			// This error is often not fatal; it can mean the backend doesn't exist yet.
-			logger.WithFields(logrus.Fields{
-				"backend": backend,
-				"error":   err,
-			}).Warning("Could not get servers for backend, it may be new. Proceeding with reconciliation.")
-			haproxyAPICallsFailed.WithLabelValues("get_servers_state").Inc()
-			statsCountVec("haproxy_api_calls_failed_total", 1, "get_servers_state")
-			// Pass an empty slice so reconciliation can proceed to add servers.
-			currentServersForBackend = []*runtime_models.RuntimeServer{}
-		} else {
-			haproxyAPICallsSuccessful.WithLabelValues("get_servers_state").Inc()
-			statsCountVec("haproxy_api_calls_successful_total", 1, "get_servers_state")
-		}
-
-		if err := reconcileHAProxyBackend(runtimeClient, backend, hosts, currentServersForBackend); err != nil {
-			reconciliationFailedBackends[backend] = true
-			logger.WithFields(logrus.Fields{
-				"backend": backend,
-				"error":   err,
-			}).Error("Failed to reconcile HAProxy backend")
-			// Continue with other backends instead of failing completely
-		} else {
-			reconciledBackends[backend] = true
-		}
-	}
-	if len(reconciliationFailedBackends) > 0 || len(reconciledBackends) == 0 {
-		resultLabel = "error"
-		if len(reconciliationFailedBackends) > 0 {
-			logger.WithField("failed_backends", reconciliationFailedBackends).Error("Failed to reconcile some HAProxy backends")
-			updateHealthForUpstreamUpdateAPI(false, errors.New("failed to reconcile some HAProxy backends: "+fmt.Sprintf("%v", reconciliationFailedBackends)).Error())
-		}
-		if len(reconciledBackends) == 0 {
-			logger.WithField("reconciled_backends", reconciledBackends).Error("Failed to reconcile any HAProxy backends")
-			updateHealthForUpstreamUpdateAPI(false, errors.New("failed to reconcile any HAProxy backends").Error())
-		}
-		return errors.New("Reconciliation failed for some or all HAProxy backends")
-	}
-
-	resultLabel = "success"
-	logger.Info("Successfully reconciled all HAProxy backends")
-	updateHealthForUpstreamUpdateAPI(true, "OK")
-
-	return nil
-}
-
-func reconcileHAProxyBackend(client runtime_api.Runtime, backend string, desiredHosts []Host, currentServers []*runtime_models.RuntimeServer) error {
-	var errs []string
-	logger.WithField("backend", backend).Debug("Reconciling HAProxy backend")
-
-	desiredServerMap := haproxyBuildDesiredServerMap(desiredHosts)
-	currentServerMap := haproxyBuildCurrentServerMap(currentServers)
-
-	if haproxyAreServerMapsIdentical(desiredServerMap, currentServerMap) {
-		logger.WithField("backend", backend).Debug("HAProxy backend is already in the desired state. No changes needed.")
-		return nil
-	}
-
-	if err := haproxyAddOrUpdateServers(client, backend, desiredServerMap, currentServerMap); err != nil {
-		errs = append(errs, fmt.Sprintf("add/update servers: %v", err))
-	}
-	//add or update servers first to avoid downtime in case of complete replacement of servers
-	if err := haproxyRemoveStaleServers(client, backend, currentServers, desiredServerMap); err != nil {
-		errs = append(errs, fmt.Sprintf("remove stale servers: %v", err))
-	}
-
-	if len(errs) > 0 {
-		return fmt.Errorf("failed to reconcile HAProxy backend: %v", errs)
-	}
-
-	logger.WithField("backend", backend).Info("Successfully reconciled HAProxy backend")
-	return nil
-}
-
-func haproxyAreServerMapsIdentical(desiredMap map[string]Host, currentMap map[string]runtime_models.RuntimeServer) bool {
-	if len(desiredMap) != len(currentMap) {
-		return false
-	}
-
-	for serverName, desiredHost := range desiredMap {
-		currentServer, exists := currentMap[serverName]
-		if !exists {
-			return false
-		}
-
-		// Resolve desired hostname to IP for a reliable comparison with HAProxy's runtime state.
-		desiredIP := desiredHost.Host // Default to hostname if resolution fails.
-		resolveCtx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
-		defer cancel()
-		if resolvedIP, err := resolveHostnameToIP(resolveCtx, desiredHost.Host); err == nil {
-			desiredIP = resolvedIP
-		}
-
-		portMatches := currentServer.Port != nil && *currentServer.Port == int64(desiredHost.Port)
-		addressMatches := currentServer.Address == desiredIP
-		adminStateMatches := currentServer.AdminState == "ready"
-		// We only check for 'up' or 'maint' as operational state can fluctuate (e.g., 'going down').
-		// A server in 'maint' is operationally down but administratively configured, so we consider it a match if we want it 'up'.
-		opStateMatches := currentServer.OperationalState == "up" || currentServer.OperationalState == "maint"
-
-		if !(portMatches && addressMatches && adminStateMatches && opStateMatches) {
-			return false // Properties do not match.
-		}
-	}
-
-	return true
-}
-
-func haproxyBuildDesiredServerMap(desiredHosts []Host) map[string]Host {
-	serverMap := make(map[string]Host)
-	for _, host := range desiredHosts {
-		serverName := generateStableHaproxyServerName(host)
-		serverMap[serverName] = host
-	}
-	return serverMap
-}
-
-func haproxyBuildCurrentServerMap(currentServers []*runtime_models.RuntimeServer) map[string]runtime_models.RuntimeServer {
-	serverMap := make(map[string]runtime_models.RuntimeServer)
-	for _, srv := range currentServers {
-		serverMap[srv.Name] = *srv
-	}
-	return serverMap
-}
-
-func haproxyRemoveStaleServers(client runtime_api.Runtime, backend string, currentServers []*runtime_models.RuntimeServer, desiredServerMap map[string]Host) error {
-	var errs []string
-	for _, srv := range currentServers {
-		if _, exists := desiredServerMap[srv.Name]; !exists {
-			logger.WithFields(logrus.Fields{"backend": backend, "server": srv.Name}).Info("Disabling and deleting stale server")
-			// Set server state to 'maint' before deleting.
-			if err := client.SetServerState(backend, srv.Name, "maint"); err != nil {
-				haproxyAPICallsFailed.WithLabelValues("set_server_state_maint").Inc()
-				statsCountVec("haproxy_api_calls_failed_total", 1, "set_server_state_maint")
-				logger.WithFields(logrus.Fields{"backend": backend, "server": srv.Name, "error": err}).Error("Failed to disable server")
-				errs = append(errs, fmt.Sprintf("disable %s: %v", srv.Name, err))
-			} else {
-				haproxyAPICallsSuccessful.WithLabelValues("set_server_state_maint").Inc()
-				statsCountVec("haproxy_api_calls_successful_total", 1, "set_server_state_maint")
-			}
-
-			if err := client.DeleteServer(backend, srv.Name); err != nil {
-				haproxyAPICallsFailed.WithLabelValues("delete_server").Inc()
-				statsCountVec("haproxy_api_calls_failed_total", 1, "delete_server")
-				logger.WithFields(logrus.Fields{"backend": backend, "server": srv.Name, "error": err}).Error("Failed to delete server")
-				errs = append(errs, fmt.Sprintf("delete %s: %v", srv.Name, err))
-			} else {
-				haproxyAPICallsSuccessful.WithLabelValues("delete_server").Inc()
-				statsCountVec("haproxy_api_calls_successful_total", 1, "delete_server")
-			}
-		}
-	}
-	if len(errs) > 0 {
-		return fmt.Errorf("Errors disabling stale servers: %s", strings.Join(errs, "; "))
-	}
-	return nil
-}
-
-func haproxyAddOrUpdateServers(client runtime_api.Runtime, backend string, desiredServerMap map[string]Host, currentServerMap map[string]runtime_models.RuntimeServer) error {
-	var errs []string
-	for serverName, host := range desiredServerMap {
-		if _, exists := currentServerMap[serverName]; !exists {
-			if err := haproxyAddNewServer(client, backend, serverName, host); err != nil {
-				errs = append(errs, fmt.Sprintf("add %s: %v", serverName, err))
-			}
-		} else {
-			if err := haproxyUpdateExistingServer(client, backend, serverName, host, currentServerMap[serverName]); err != nil {
-				errs = append(errs, fmt.Sprintf("update %s: %v", serverName, err))
-			}
-		}
-	}
-	if len(errs) > 0 {
-		return fmt.Errorf("Errors in Add/Update servers: %s", strings.Join(errs, "; "))
-	}
-	return nil
-}
-
-func haproxyAddNewServer(client runtime_api.Runtime, backend, serverName string, host Host) error {
-	logger.WithFields(logrus.Fields{"backend": backend, "server": serverName}).Info("Adding new server")
-	if err := client.AddServer(backend, serverName, fmt.Sprintf("%s:%d", host.Host, host.Port)); err != nil {
-		haproxyAPICallsFailed.WithLabelValues("add_server").Inc()
-		statsCountVec("haproxy_api_calls_failed_total", 1, "add_server")
-		logger.WithFields(logrus.Fields{"backend": backend, "server": serverName, "error": err}).Error("Failed to add server")
-		return err
-	}
-	haproxyAPICallsSuccessful.WithLabelValues("add_server").Inc()
-	statsCountVec("haproxy_api_calls_successful_total", 1, "add_server")
-
-	if err := client.SetServerState(backend, serverName, "ready"); err != nil {
-		haproxyAPICallsFailed.WithLabelValues("set_server_state_ready").Inc()
-		statsCountVec("haproxy_api_calls_failed_total", 1, "set_server_state_ready")
-		logger.WithFields(logrus.Fields{"backend": backend, "server": serverName, "error": err}).Error("Failed to set new server state to ready")
-		return err
-	}
-	haproxyAPICallsSuccessful.WithLabelValues("set_server_state_ready").Inc()
-	statsCountVec("haproxy_api_calls_successful_total", 1, "set_server_state_ready")
-	return nil
-}
-
-func haproxyUpdateExistingServer(client runtime_api.Runtime, backend, serverName string, host Host, currentServer runtime_models.RuntimeServer) error {
-	// Resolve desired hostname to IP for a comparison with HAProxy's runtime state.
-	desiredIP := host.Host // Default to hostname if resolution fails. This will help a force update even if resolution is unavailable.
-	resolveCtx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
-	defer cancel()
-	if resolvedIP, err := resolveHostnameToIP(resolveCtx, host.Host); err != nil {
-		logger.WithFields(logrus.Fields{"hostname": host.Host, "error": err}).Warning("Failed to resolve hostname; using hostname for comparison")
-	} else {
-		desiredIP = resolvedIP
-	}
-
-	portMatches := currentServer.Port != nil && *currentServer.Port == int64(host.Port)
-	if currentServer.Address == desiredIP && portMatches && currentServer.AdminState == "ready" && currentServer.OperationalState == "up" {
-		logger.WithFields(logrus.Fields{"backend": backend, "server": serverName}).Trace("Server is up-to-date, no update needed")
-		return nil
-	}
-
-	logger.WithFields(logrus.Fields{
-		"backend":           backend,
-		"server":            serverName,
-		"current_address":   currentServer.Address,
-		"current_port":      *currentServer.Port,
-		"current_admin":     currentServer.AdminState,
-		"current_operstate": currentServer.OperationalState,
-		"desired_address":   desiredIP,
-		"desired_port":      host.Port,
-	}).Info("Updating existing server")
-
-	var errs []string
-
-	if err := client.SetServerAddr(backend, serverName, desiredIP, int(host.Port)); err != nil {
-		haproxyAPICallsFailed.WithLabelValues("set_server_addr").Inc()
-		statsCountVec("haproxy_api_calls_failed_total", 1, "set_server_addr")
-		logger.WithFields(logrus.Fields{"backend": backend, "server": serverName, "error": err}).Error("Failed to set server address")
-		errs = append(errs, fmt.Sprintf("set address: %v", err))
-	} else {
-		haproxyAPICallsSuccessful.WithLabelValues("set_server_addr").Inc()
-		statsCountVec("haproxy_api_calls_successful_total", 1, "set_server_addr")
-	}
-
-	if err := client.SetServerState(backend, serverName, "ready"); err != nil {
-		haproxyAPICallsFailed.WithLabelValues("set_server_state_ready").Inc()
-		statsCountVec("haproxy_api_calls_failed_total", 1, "set_server_state_ready")
-		logger.WithFields(logrus.Fields{"backend": backend, "server": serverName, "error": err}).Error("Failed to set server state to ready")
-		errs = append(errs, fmt.Sprintf("set state: %v", err))
-	} else {
-		haproxyAPICallsSuccessful.WithLabelValues("set_server_state_ready").Inc()
-		statsCountVec("haproxy_api_calls_successful_total", 1, "set_server_state_ready")
-	}
-
-	if len(errs) > 0 {
-		return fmt.Errorf("haproxyUpdateExistingServer errors: %s", strings.Join(errs, "; "))
-	}
-	return nil
 }
 
 func resolveHostnameToIP(ctx context.Context, hostname string) (string, error) {
@@ -1163,4 +776,3 @@ func reloadWorker() {
 		}
 	}()
 }
-
